Move extension to common
