+++ src/main/State.ts
@@ -1,5 +1,6 @@
+import Extension from '../common/Extension';
+
 import Bookshelf from './Bookshelf';
-import Extension from './Extension';
 
 export interface State {
   bookshelf: Bookshelf;
